requirements.txt
Django==4.2.1
Pillow==9.5.0
Pillow==10.2.0
rembg==2.0.56
